Integrating the latest changes from the MATE21 repo. The program now compensates for the camera's intrinsic factors (kinect IR sensor parameters are currently hardcoded in). The reconstruction function now has parameters to set the min and max capture distances of your depth capture, we need this to estimate the distance in mm from the depth data.

+++ Depth2Mesh/MeshWriter.cpp
@@ -10,7 +10,7 @@ MeshWriter::~MeshWriter()
 {
 }
 
-float MeshWriter::interpolate(float a, float b, float coefficient)
+double MeshWriter::interpolate(double a, double b, double coefficient)
 {
 	return a + coefficient * (b - a);
 }
